修改个人页面的余额显示格式，使其支持小数点后两位

+++ living/living/Model/VO/UserInfo/UserInfoVO.h
@@ -28,7 +28,7 @@
 @property (nonatomic, strong) NSString *city;
 @property (nonatomic, strong) NSString *avatar;
 @property (nonatomic, strong) NSString *eventNum;
-@property (nonatomic, assign) int balance;
+@property (nonatomic, assign) float balance;
 @property (nonatomic, strong) NSString *birthday;
 @property (nonatomic, assign) int orderNumber;
 @property (nonatomic, strong) NSString *totalEventNum;
